+++ src/AbstractMicroplate.php
@@ -4,7 +4,6 @@ namespace Mll\Microplate;
 
 use Illuminate\Support\Collection;
 use Mll\Microplate\Enums\FlowDirection;
-use Mll\Microplate\Exceptions\MicroplateIsFullException;
 use Mll\Microplate\Exceptions\UnexpectedFlowDirection;
 
 /**
@@ -23,27 +22,17 @@ abstract class AbstractMicroplate
     public CoordinateSystem $coordinateSystem;
 
     /**
-     * @var WellsCollection
-     */
-    protected Collection $wells;
-
-    /**
      * @param TCoordinateSystem $coordinateSystem
      */
     public function __construct(CoordinateSystem $coordinateSystem)
     {
         $this->coordinateSystem = $coordinateSystem;
-
-        $this->clearWells();
     }
 
     /**
      * @return WellsCollection
      */
-    public function wells(): Collection
-    {
-        return $this->wells;
-    }
+    abstract public function wells(): Collection;
 
     /**
      * @param Coordinate<TCoordinateSystem> $coordinate
@@ -95,7 +84,7 @@ abstract class AbstractMicroplate
      */
     public function freeWells(): Collection
     {
-        return $this->wells->filter(
+        return $this->wells()->filter(
             /**
              * @param TWell $value
              */
@@ -114,54 +103,5 @@ abstract class AbstractMicroplate
              */
             static fn ($value): bool => self::EMPTY_WELL !== $value
         );
-    }
-
-    /**
-     * Clearing the wells will reinitialize all well position of the coordinate system.
-     */
-    public function clearWells(): void
-    {
-        /**
-         * Flow direction is irrelevant during initialization, it is not a property of
-         * a plate but rather a property of the access to the plate.
-         */
-        $this->wells = new Collection();
-        foreach ($this->coordinateSystem->columns() as $column) {
-            foreach ($this->coordinateSystem->rows() as $row) {
-                $this->wells[$row . $column] = self::EMPTY_WELL;
-            }
-        }
-    }
-
-    /**
-     * @throws MicroplateIsFullException
-     *
-     * @return Coordinate<TCoordinateSystem>
-     */
-    public function nextFreeWellCoordinate(FlowDirection $flowDirection): Coordinate
-    {
-        $coordinateString = $this->sortedWells($flowDirection)
-            ->search(self::EMPTY_WELL);
-
-        if (! is_string($coordinateString)) {
-            throw new MicroplateIsFullException();
-        }
-
-        return Coordinate::fromString($coordinateString, $this->coordinateSystem);
-    }
-
-    /**
-     * @param TWell $content
-     *
-     * @throws MicroplateIsFullException
-     *
-     * @return Coordinate<TCoordinateSystem>
-     */
-    public function addToNextFreeWell($content, FlowDirection $flowDirection): Coordinate
-    {
-        $coordinate = $this->nextFreeWellCoordinate($flowDirection);
-        $this->wells[$coordinate->toString()] = $content;
-
-        return $coordinate;
     }
 }